Transition all dependency injection to array notation to avoid errors with minification. Minified scripts translate variables like function(Dependency) to function(a). Since these variable names are used to lookup dependencies in the system, renamed variables won't be looked up appropriately.
Array notation uses ['Dependency', function(Dependency) {}]; The names of the dependencies are looked up in the array, and then passed in in order as arguments. Because they are now ordered arguments, it's okay for them to be renamed, since the array itself own't be minified (the data would still be important and survives minification). Now we can minify the script for the build.